feat: 🎸 Remove redundant check auth
